refactor: streamline editor initialization and update handling in NoteViewer and Rich_Editor components
